src/mpass/test_requirements.txt
# The test_requirements.txt should contain everything that is needed in order to
# get the tests of this app up and running.
#
# Tests can then be executed simply by running ``tox``

django==1.4
factory_boy==1.2.0
django-nose==1.1
coverage==3.0
django-coverage
mock==1.0.1
selenium

# Any internal library dependencies must be listed also, for example:
# -e hg+ssh://hg@bitbucket.org/haltu/pypi-haltu-utils@v2.6.0#egg=haltu-utils

